requirements.txt
torch>=2.1.0
transformers>=4.40.0
peft>=0.10.0
datasets>=2.18.0
accelerate>=0.28.0
bitsandbytes>=0.43.0
trl>=0.8.0
PyGithub>=2.3.0
pydantic>=2.6.0
lm-format-enforcer>=0.10.3
jsonschema>=4.21.0
openai>=1.14.0
tiktoken>=0.6.0
rouge-score>=0.1.2
numpy>=1.26.0
pandas>=2.2.0
tqdm>=4.66.0
rich>=13.7.0
PyYAML>=6.0.1
python-dotenv>=1.0.1
